detector qrs em C no modelo simulink

+++ matlab/realtime/init_model.m
@@ -1,6 +1,6 @@
 %% ECG data
 signal = utils.interpret_ecg(e0103, 1);
-data = signal.data - signal.inival;
+data = int16(signal.data - signal.inival);
 simin = timeseries(data, signal.time);
 Fs = signal.fs;
 Fm = 50;
